Use "unsigned char:1" for flags
"char" is treated, at least in many cases, as signed values. Having
that field as 1 bit means it can be treated as one sign bit, no
data bits. Thus only values 0 and -1 will be accepted

It seems to differ between versions of the compiler. But in any
case, explicitly defining it as unsigned means only the values 0
and 1 will be accepted, and therefore there shouldn't be any issues
between compiler versions